move mnist dataset to another folder
