+++ tests/_tests_ml_cmcscore/experiment.py
@@ -3,7 +3,7 @@ from collections import OrderedDict
 
 from catalyst.contrib.data.transforms import Compose, Normalize, ToTensor
 from catalyst.contrib.datasets import MNIST
-from catalyst.contrib.datasets.metric_learning import MnistQGDataset
+from catalyst.contrib.datasets.mnist import MnistQGDataset
 from catalyst.dl import ConfigExperiment
